Don't clean up URLs; try sending source in query string.
Something looks fishy about the cross-origin referrer header coming from
iOS-based browsers, despite the `unsafe-url` policy. Removing the URL
clean-up and therefore the manual setting of the fetch `referrer`. Also,
test putting the location in the query string, too---it'll be interesting
to see if there are discrepancies between it and the HTTP `Referer`.